feat(raw): native raw import on the web build, LibRaw prompt on the desktop
The raw codec was gated off wasm along with LibRaw, though the native
decoder it runs first is pure Rust. The LibRaw fallback now lives in its
own off-wasm module, the codec registers on every target, and a browser
refuses the few files only LibRaw develops (Canon CR3 pixels, Fuji
compressed RAF, some odd codecs) with a message saying it is the web
build; their embedded previews still show.

On the desktop, opening one of those files on a machine with no LibRaw
now raises a dialog that says which library is needed and how to
install it for the platform, with a retry, instead of a bare status
error.
