+++ crates/app/src/workspace/modals.rs
@@ -466,6 +466,7 @@ impl Workspace {
             | Modal::DropImage { .. }
             | Modal::DropFolders { .. }
             | Modal::HeifSupport { .. }
+            | Modal::LibRawSupport { .. }
             | Modal::CameraImport { .. }
             | Modal::CameraImportOptions { .. }
             | Modal::CameraImportFailed { .. }
